suppression de notion de bouton  et de selection unique, lancement du lien au clic sur un élément du slider + css + ajout d'images pour chaque liste et son contenu
