+++ src/App.js
@@ -1,5 +1,4 @@
 import React from "react";
-import logo from "./logo.svg";
 import Home from './pages/home';
 
 function App() {
