+++ src/styles/movielists.css
@@ -14,14 +14,14 @@
     margin-block-end: 0px;
     margin: 50px 0 50px 0;
 }
-.list-category {
+.list-container .list-category {
     text-transform:uppercase;
     font-weight:700;
     font-size:15px;
     position:relative;
     margin-block-end: 20px;
 }
-.list-category::after {
+.list-container .list-category::after {
     content:'';
     width:calc(100% - 92px);
     height:1px;
@@ -33,72 +33,76 @@
 }
 
 /* CAROUSEL */
-.carousel-inner {
+.list-container .carousel-inner {
     width:70%!important;
     margin:auto;
+    border-radius: 15px;
 }
-.carousel-control-prev, .carousel-control-next {
-    background-color:rgba(102, 198, 155, .7);
+.list-container .carousel-control-prev,.list-container .carousel-control-next {
+    background-color:rgba(102, 198, 155, 1);
     height: 80%;
     top: 50%!important;
     transform: translateY(-50%);
     width: 12%!important;
 }
-.carousel-control-prev {
+.list-container .carousel-control-prev {
     border-radius: 10px 0px 0px 10px;
 }
-.carousel-control-next {
+.list-container .carousel-control-next {
     border-radius: 0px 10px 10px 0px;
 }
-.carousel-item {
+.list-container .carousel-item {
     height: 135px;
     position:relative;
-    border-radius: 15px;
+    border-radius: 10px;
     cursor: pointer;
-    transition:all 0.2s ease-in-out;
+    overflow: hidden;
+    border: 3px solid white;
 }
-.carousel-item:after {
-    position:absolute;
-    top:0;
-    left:0;
-    height:100%;
-    width:100%;
-    content:'';
-    background-image:linear-gradient(45deg, rgba(0, 0, 0, 0) 60%, rgba(239, 239, 239, 0.3) 80%);
-    background-size:150%;
-    background-repeat:repeat;
-    transition:all 0.2s ease-in-out;
-}
-.carousel-item:hover:after {
+.list-container .carousel-item:hover {
     background-size:100%;
+    border:3px solid #66C69B;
 }
-/*.carousel-item::after {
-    position:absolute;
-    content:'';
-    width:100%;
-    height:100%;
-    background-image:linear-gradient(0deg, rgba(0, 0, 0, 0.8) 22.22%, rgba(128, 128, 128, 0.5) 75.69%, rgba(255, 255, 255, 0) 100%);
-    top:0;
-}*/
-.carousel-caption h3 {
+.list-container .carousel-caption {
+    text-shadow: 0 0 3px black;
+}
+.list-container .carousel-caption h3 {
     margin-block-end:0px;
-    line-height: 22px;
+    text-shadow: 0 0 3px black;
+    text-transform: uppercase;
+    font-weight: bold;
+    font-size: 20px;
 }
-.carousel-caption.d-none {
-    display:block!important;
+.list-container .carousel-caption p {
+    margin-block-end:0px;
+    font-weight: lighter;
+    font-size: 16px;
 }
-.clickable-list {
+.list-container .carousel-caption.d-none {
+    z-index: 9999999;
+    display: flex!important;
+    align-items: center;
+    justify-content: center;
+    height: 100%;
+    top: 0;
+    width: 100%;
+    left: 0;
+    flex-direction: column;
+}
+.list-container .clickable-list {
     position:absolute;
     height:100%;
     width:100%;
     z-index:999999;
     transition:all 0.2s ease-in-out;
-    background:linear-gradient(to right, rgba(102, 198, 155, 0), rgba(102, 198, 155, 0))
+    background-repeat:no-repeat;
+    background-size:cover;
+    background-position:center;
 }
-.clickable-list.selected-list {
+.list-container .clickable-list.selected-list {
     background: linear-gradient(to right, rgba(102, 198, 155, 0.7), rgba(102, 198, 155, 0.7));
 }
-.carousel-item img {
+.list-container .carousel-item img {
     position: absolute;
-    width: 100%;
+    height: 100%;
 }
